Remove DayOfWeekEnum and related references
This commit removes the `DayOfWeekEnum` and its associated `Enum` property from the `DayOfWeekEntity` class. All references to the enum have been eliminated from the `DayOfWeekEntityConfiguration` and `ApplicationDbContext`, including the unique index configuration and data seeding. The changes reflect a shift away from using an enumeration for days of the week in the application.